modelserver-fastapi/requirements.txt
fastapi
pydantic
transformers
uvicorn
scikit-learn
python-dotenv
mlflow
pandas
asyncio
httpx
langchain
chromadb==0.3.29
langchain_community
pypdf
google-generativeai
google-adk
